[MUYAHO-6] 멤버가 보유한 비트코인의 현재가를 조회해서 계산하는 기능 추가. (#28)

+++ muyaho-api/src/main/java/com/depromeet/muyaho/service/memberstock/dto/request/AddMemberStockRequest.java
@@ -27,7 +27,7 @@ public class AddMemberStockRequest {
         return MemberStock.of(memberId, stock, purchasePrice, quantity);
     }
 
-    public static AddMemberStockRequest testInstance(Long stockId, int purchasePrice, int quantity) {
+    public static AddMemberStockRequest testInstance(Long stockId, double purchasePrice, double quantity) {
         return new AddMemberStockRequest(stockId, purchasePrice, quantity);
     }
